moves dto into adapters
